refactor(controller package): move createPost and createPromoPost from PostController to ProductController, and delete PostController
